Preserve IV in masked passwords
For block ciphers that are chained Cipher implementations, if not given
any, automatically pick a random Initialization Vector (IV), this needs
to be preserved in order for the decryption to succeed.

This does that by prepending the IV to the resulting cipher, extracts
and passes it in IvParameterSpec in the reverse. It does away with the
`getCipher` method as the symmetry between `mask` and `unmask` is no
longer present.

The test presented in `MaskedPasswordImplTest` fails for PBKDF*
algorithms as they are not supported `Cipher` algorithms.