PicketBox compatibility
PicketBox did not preserve the IV, as it happens it used (fixed)
`PBEwithMD5andDES` in `PBES1` mode (RFC2898) for which IV is derived
from secret key.

Thats why `PBEwithMD5andDES` needs to be handled differently to make the
implementation compatible with PicketBox masked passwords.

This commit adds a conditions that for `PBEwithMD5andDES` sets the block
size to 0 and does not prepend the IV to the resulting masked password
bytes.