[ELY-2326] Update GSSCredential and KerberosTicket validity checking.
Both GSSCredentialSecurityFactory implementations are equivalent, the
one in util is deprecated to be removed.

Updated to use an AtomicReference along with a UnaryOperator to update
the cached GSSKerberosCredential as needed.

Added checking of the validity of the KerberosTicket and an attempt to
renew directly.  If all else fails obtain new credentials.

New logging added that would have helped diagnose the issue.