[ELY-2360] Change OIDC_STATE delimiter
Change the delimiter character sequence used in the OIDC_STATE cookie
to avoid clashes when the same character sequence appears in the encoded
tokens stored in the cookie.

Additionally, a debug log message is added to print the value of the
OIDC_STATE cookie if the value could not be split into its components.